+++ actix-web-lab/src/lib.rs
@@ -43,6 +43,7 @@ mod lazy_data;
 mod load_shed;
 mod local_data;
 mod middleware_from_fn;
+mod middleware_map_response;
 mod ndjson;
 mod normalize_path;
 mod panic_reporter;